refactor(ui): factor avatar initials and color helpers into s/avatarUtils
Removes the duplicated initials/hash helpers copy-pasted in pipeline,
orgManager and metadataRetriever (jscpd clones), and declares the CSS
font-stack names Segoe and Neue in the cspell dictionary.
