+++ .github/linters/.cspell.json
@@ -96,6 +96,7 @@
     "SLDS",
     "Schedulable",
     "Scrolldown",
+    "Segoe",
     "Gitbeaker",
     "azdev",
     "busalesforce",
@@ -338,6 +339,7 @@
     "minimalpermsets",
     "Italiano",
     "Nederlands",
+    "Neue",
     "licensetypes",
     "returnactiveusers",
     "instanceupgrade",
